Add autorefresh to accounts
Add an additional web3 instance that uses infura's mainnet websocket
endpoint.  On every new block we simply refresh the balance,
transactions and unique tokens for a given account.

Few things to note:
- ACCOUNT_UPDATE_BALANCES_REQUEST wasn't being used so, removed it.
- accountClearIntervals looks like it was replaced in favor of
accountClearState, so removed that function as well.
- web3 v1.0.0-beta34 that we're using has a bug connecting to any
of the testnets on infura, downgrading to web3 v1.0.0-beta33 fixes
the problem if we want to have a network specific subscription
https://github.com/ethereum/web3.js/issues/1559